Working on Alphabet bug
